openfin-backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
tinyfish
pyjwt
